requirements.txt
cryptography==1.7.1
daemonize==2.4.7
idna==2.2
keyring==10.1
keyrings.alt==1.3
pyasn1==0.1.9
pycrypto==2.6.1
PyYAML==3.12
